refactor source tree: libgit2 exports stay in libgit2
The functions exported by libgit2 should stay in the libgit2 directory.
By putting them in `util`, they'll be further exported by any other tool
that uses the util library.
